+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added wizard warning about google passwords. [sourcehut#41]
+
 ### Changed
 
 - Changed `envelope.watch.{event}.{hook}`: hooks can now be cumulated. For example it is possible to send a system notification and execute a shell command when receiving a new envelope:
@@ -900,6 +904,7 @@ Few major concepts changed:
 [#419]: https://github.com/soywod/himalaya/issues/419
 [#430]: https://github.com/soywod/himalaya/issues/430
 
+[sourcehut#41]: https://todo.sr.ht/~soywod/pimalaya/41
 [sourcehut#43]: https://todo.sr.ht/~soywod/pimalaya/43
 [sourcehut#54]: https://todo.sr.ht/~soywod/pimalaya/54
 [sourcehut#59]: https://todo.sr.ht/~soywod/pimalaya/59